Add files via upload
My main aim was to tidy up the start and log in on my PClinuxOS 64 bit plasma 5 desktop. I like the Hand scan splash screen and starting with the old version that had been ported for plasma 5. I found it had many problems and would not scale not all the graphics were even being displayed .So after a lot of reading and learning I have rewritten the QML file that displays the splash screen as well as updated some of the graphics, to create a much more dynamic and I think interesting thing to look at while my machine continues its journey towards logging in. I also found the Air desktop causes a bug where widgets cannot be deleted from the desktop without locking the desktop up so this version of Oxygen uses Breeze.